@@ -1,7 +1,7 @@
 [build-system]
 requires = [
     "setuptools>=62.6,<=80.9.0",
-    "setuptools-scm>=6.2,<=9.2.0",
+    "setuptools-scm>=9.2,<=9.2.2",
 ]
 build-backend = "setuptools.build_meta"
 
@@ -12,7 +12,7 @@ authors = [
 ]
 description = "A Python wrapper for the Discord API"
 readme = {content-type = "text/x-rst", file = "README.rst"}
-requires-python = ">=3.9"
+requires-python = ">=3.10, <3.14"
 license = "MIT"
 license-files = ["LICENSE"]
 classifiers = [
@@ -20,7 +20,6 @@ classifiers = [
     "Intended Audience :: Developers",
     "Natural Language :: English",
     "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -68,7 +67,7 @@ voice = {file = "requirements/voice.txt"}
 [tool.setuptools_scm]
 
 [tool.black]
-target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
+target-version = ['py310', 'py311', 'py312', 'py313']
 
 [tool.isort]
 profile = "black"
@@ -87,7 +86,7 @@ extension-pkg-whitelist = [
     "pydantic",
     "ujson"
 ]
-py-version = 3.9
+py-version = 3.10
 
 [tool.pylint.messages_control]
 enable = [
