@@ -26,9 +26,25 @@ from __future__ import annotations
 
 import inspect
 import logging
+import sys
 import types
+from collections.abc import Awaitable, Callable, Iterable
 from enum import Enum
-from typing import TYPE_CHECKING, Literal, Optional, Type, Union, get_args
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Literal,
+    Optional,
+    Type,
+    TypeVar,
+    Union,
+    get_args,
+)
+
+if sys.version_info >= (3, 12):
+    from typing import TypeAliasType
+else:
+    from typing_extensions import TypeAliasType
 
 from ..abc import GuildChannel, Mentionable
 from ..channel import (
@@ -41,13 +57,14 @@ from ..channel import (
     Thread,
     VoiceChannel,
 )
-from ..commands import ApplicationContext
+from ..commands import ApplicationContext, AutocompleteContext
 from ..enums import ChannelType
 from ..enums import Enum as DiscordEnum
 from ..enums import SlashCommandOptionType
 from ..utils import MISSING, basic_autocomplete
 
 if TYPE_CHECKING:
+    from ..cog import Cog
     from ..ext.commands import Converter
     from ..member import Member
     from ..message import Attachment
@@ -72,6 +89,25 @@ if TYPE_CHECKING:
         Type[Enum],
         Type[DiscordEnum],
     ]
+
+    AutocompleteReturnType = Union[
+        Iterable["OptionChoice"], Iterable[str], Iterable[int], Iterable[float]
+    ]
+    T = TypeVar("T", bound=AutocompleteReturnType)
+    MaybeAwaitable = Union[T, Awaitable[T]]
+    AutocompleteFunction = Union[
+        Callable[[AutocompleteContext], MaybeAwaitable[AutocompleteReturnType]],
+        Callable[[Cog, AutocompleteContext], MaybeAwaitable[AutocompleteReturnType]],
+        Callable[
+            [AutocompleteContext, Any],  # pyright: ignore [reportExplicitAny]
+            MaybeAwaitable[AutocompleteReturnType],
+        ],
+        Callable[
+            [Cog, AutocompleteContext, Any],  # pyright: ignore [reportExplicitAny]
+            MaybeAwaitable[AutocompleteReturnType],
+        ],
+    ]
+
 
 __all__ = (
     "ThreadOption",
@@ -149,15 +185,6 @@ class Option:
     max_length: Optional[:class:`int`]
         The maximum length of the string that can be entered. Must be between 1 and 6000 (inclusive).
         Only applies to Options with an :attr:`input_type` of :class:`str`.
-    autocomplete: Optional[Callable[[:class:`.AutocompleteContext`], Awaitable[Union[Iterable[:class:`.OptionChoice`], Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]]]]]
-        The autocomplete handler for the option. Accepts a callable (sync or async)
-        that takes a single argument of :class:`AutocompleteContext`.
-        The callable must return an iterable of :class:`str` or :class:`OptionChoice`.
-        Alternatively, :func:`discord.utils.basic_autocomplete` may be used in place of the callable.
-
-        .. note::
-
-            Does not validate the input value against the autocomplete results.
     channel_types: list[:class:`discord.ChannelType`] | None
         A list of channel types that can be selected in this option.
         Only applies to Options with an :attr:`input_type` of :class:`discord.SlashCommandOptionType.channel`.
@@ -197,6 +224,7 @@ class Option:
         if self.name is not None:
             self.name = str(self.name)
         self._parameter_name = self.name  # default
+        input_type = self._parse_type_alias(input_type)
         input_type = self._strip_none_type(input_type)
         self._raw_type: InputType | tuple = input_type
 
@@ -278,6 +306,7 @@ class Option:
         )
         self.default = kwargs.pop("default", None)
 
+        self._autocomplete: AutocompleteFunction | None = None
         self.autocomplete = kwargs.pop("autocomplete", None)
         if len(enum_choices) > 25:
             self.choices: list[OptionChoice] = []
@@ -368,6 +397,12 @@ class Option:
             raise TypeError("input_type cannot be NoneType.")
 
     @staticmethod
+    def _parse_type_alias(input_type: InputType) -> InputType:
+        if isinstance(input_type, TypeAliasType):
+            return input_type.__value__
+        return input_type
+
+    @staticmethod
     def _strip_none_type(input_type):
         if isinstance(input_type, SlashCommandOptionType):
             return input_type
@@ -422,6 +457,43 @@ class Option:
 
     def __repr__(self):
         return f"<discord.commands.{self.__class__.__name__} name={self.name}>"
+
+    @property
+    def autocomplete(self) -> AutocompleteFunction | None:
+        """
+        The autocomplete handler for the option. Accepts a callable (sync or async)
+        that takes a single required argument of :class:`AutocompleteContext` or two arguments
+        of :class:`discord.Cog` (being the command's cog) and :class:`AutocompleteContext`.
+        The callable must return an iterable of :class:`str` or :class:`OptionChoice`.
+        Alternatively, :func:`discord.utils.basic_autocomplete` may be used in place of the callable.
+
+        Returns
+        -------
+        Optional[AutocompleteFunction]
+
+        .. versionchanged:: 2.7
+
+        .. note::
+            Does not validate the input value against the autocomplete results.
+        """
+        return self._autocomplete
+
+    @autocomplete.setter
+    def autocomplete(self, value: AutocompleteFunction | None) -> None:
+        self._autocomplete = value
+        # this is done here so it does not have to be computed every time the autocomplete is invoked
+        if self._autocomplete is not None:
+            self._autocomplete._is_instance_method = (  # pyright: ignore [reportFunctionMemberAccess]
+                sum(
+                    1
+                    for param in inspect.signature(
+                        self._autocomplete
+                    ).parameters.values()
+                    if param.default == param.empty  # pyright: ignore[reportAny]
+                    and param.kind not in (param.VAR_POSITIONAL, param.VAR_KEYWORD)
+                )
+                == 2
+            )
 
 
 class OptionChoice:
