@@ -1,3 +1,27 @@
+"""
+The MIT License (MIT)
+
+Copyright (c) 2021-present Pycord Development
+
+Permission is hereby granted, free of charge, to any person obtaining a
+copy of this software and associated documentation files (the "Software"),
+to deal in the Software without restriction, including without limitation
+the rights to use, copy, modify, merge, publish, distribute, sublicense,
+and/or sell copies of the Software, and to permit persons to whom the
+Software is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+DEALINGS IN THE SOFTWARE.
+"""
+
 from __future__ import annotations
 
 import os
@@ -5,6 +29,7 @@ from typing import TYPE_CHECKING
 
 from ..components import InputText as InputTextComponent
 from ..enums import ComponentType, InputTextStyle
+from .item import ModalItem
 
 __all__ = ("InputText", "TextInput")
 
@@ -13,7 +38,7 @@ if TYPE_CHECKING:
     from ..types.components import InputText as InputTextComponentPayload
 
 
-class InputText:
+class InputText(ModalItem):
     """Represents a UI text input field.
 
     .. versionadded:: 2.0
@@ -27,11 +52,6 @@ class InputText:
     label: :class:`str`
         The label for the input text field.
         Must be 45 characters or fewer.
-    description: Optional[:class:`str`]
-        The description for the input text field.
-        Must be 100 characters or fewer.
-
-        .. versionadded:: 2.7
     placeholder: Optional[:class:`str`]
         The placeholder text that is shown if nothing is selected, if any.
         Must be 100 characters or fewer.
@@ -64,7 +84,6 @@ class InputText:
         "max_length",
         "custom_id",
         "id",
-        "description",
     )
 
     def __init__(
@@ -72,7 +91,7 @@ class InputText:
         *,
         style: InputTextStyle = InputTextStyle.short,
         custom_id: str | None = None,
-        label: str,
+        label: str | None = None,
         placeholder: str | None = None,
         min_length: int | None = None,
         max_length: int | None = None,
@@ -80,13 +99,10 @@ class InputText:
         value: str | None = None,
         row: int | None = None,
         id: int | None = None,
-        description: str | None = None,
     ):
         super().__init__()
-        if len(str(label)) > 45:
+        if label and len(str(label)) > 45:
             raise ValueError("label must be 45 characters or fewer")
-        if description and len(description) > 100:
-            raise ValueError("description must be 100 characters or fewer")
         if min_length and (min_length < 0 or min_length > 4000):
             raise ValueError("min_length must be between 0 and 4000")
         if max_length and (max_length < 0 or max_length > 4000):
@@ -100,7 +116,6 @@ class InputText:
                 f"expected custom_id to be str, not {custom_id.__class__.__name__}"
             )
         custom_id = os.urandom(16).hex() if custom_id is None else custom_id
-        self.description: str | None = description
 
         self._underlying = InputTextComponent._raw_construct(
             type=ComponentType.input_text,
@@ -125,18 +140,9 @@ class InputText:
         return f"<{self.__class__.__name__} {attrs}>"
 
     @property
-    def type(self) -> ComponentType:
-        return self._underlying.type
-
-    @property
     def style(self) -> InputTextStyle:
         """The style of the input text field."""
         return self._underlying.style
-
-    @property
-    def id(self) -> int | None:
-        """The input text's ID. If not provided by the user, it is set sequentially by Discord."""
-        return self._underlying.id
 
     @style.setter
     def style(self, value: InputTextStyle):
@@ -243,16 +249,15 @@ class InputText:
         return 5
 
     def to_component_dict(self) -> InputTextComponentPayload:
-        return self._underlying.to_dict()
+        return super().to_component_dict()
 
     def refresh_state(self, data) -> None:
         self._input_value = data.get("value", None)
 
-    def refresh_from_modal(self, interaction: Interaction, data: dict) -> None:
+    def refresh_from_modal(
+        self, interaction: Interaction, data: InputTextComponentPayload
+    ) -> None:
         return self.refresh_state(data)
-
-    def uses_label(self) -> bool:
-        return self.description is not None
 
 
 TextInput = InputText
