@@ -1,3 +1,27 @@
+"""
+The MIT License (MIT)
+
+Copyright (c) 2021-present Pycord Development
+
+Permission is hereby granted, free of charge, to any person obtaining a
+copy of this software and associated documentation files (the "Software"),
+to deal in the Software without restriction, including without limitation
+the rights to use, copy, modify, merge, publish, distribute, sublicense,
+and/or sell copies of the Software, and to permit persons to whom the
+Software is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+DEALINGS IN THE SOFTWARE.
+"""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, TypeVar
@@ -5,7 +29,7 @@ from typing import TYPE_CHECKING, TypeVar
 from ..components import MediaGallery as MediaGalleryComponent
 from ..components import MediaGalleryItem
 from ..enums import ComponentType
-from .item import Item
+from .item import ViewItem
 
 __all__ = ("MediaGallery",)
 
@@ -13,14 +37,14 @@ if TYPE_CHECKING:
     from typing_extensions import Self
 
     from ..types.components import MediaGalleryComponent as MediaGalleryComponentPayload
-    from .view import View
+    from .view import DesignerView
 
 
 M = TypeVar("M", bound="MediaGallery")
-V = TypeVar("V", bound="View", covariant=True)
+V = TypeVar("V", bound="DesignerView", covariant=True)
 
 
-class MediaGallery(Item[V]):
+class MediaGallery(ViewItem[V]):
     """Represents a UI Media Gallery. Galleries may contain up to 10 :class:`MediaGalleryItem` objects.
 
     .. versionadded:: 2.7
@@ -80,7 +104,7 @@ class MediaGallery(Item[V]):
         *,
         description: str = None,
         spoiler: bool = False,
-    ) -> None:
+    ) -> Self:
         """Adds a new media item to the gallery.
 
         Parameters
@@ -105,20 +129,8 @@ class MediaGallery(Item[V]):
 
         return self.append_item(item)
 
-    @Item.view.setter
-    def view(self, value):
-        self._view = value
-
-    @property
-    def type(self) -> ComponentType:
-        return self._underlying.type
-
-    @property
-    def width(self) -> int:
-        return 5
-
     def to_component_dict(self) -> MediaGalleryComponentPayload:
-        return self._underlying.to_dict()
+        return super().to_component_dict()
 
     @classmethod
     def from_component(cls: type[M], component: MediaGalleryComponent) -> M:
