@@ -1,3 +1,27 @@
+"""
+The MIT License (MIT)
+
+Copyright (c) 2021-present Pycord Development
+
+Permission is hereby granted, free of charge, to any person obtaining a
+copy of this software and associated documentation files (the "Software"),
+to deal in the Software without restriction, including without limitation
+the rights to use, copy, modify, merge, publish, distribute, sublicense,
+and/or sell copies of the Software, and to permit persons to whom the
+Software is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+DEALINGS IN THE SOFTWARE.
+"""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, TypeVar
@@ -5,20 +29,20 @@ from typing import TYPE_CHECKING, TypeVar
 from ..components import Separator as SeparatorComponent
 from ..components import _component_factory
 from ..enums import ComponentType, SeparatorSpacingSize
-from .item import Item
+from .item import ViewItem
 
 __all__ = ("Separator",)
 
 if TYPE_CHECKING:
     from ..types.components import SeparatorComponent as SeparatorComponentPayload
-    from .view import View
+    from .view import DesignerView
 
 
 S = TypeVar("S", bound="Separator")
-V = TypeVar("V", bound="View", covariant=True)
+V = TypeVar("V", bound="DesignerView", covariant=True)
 
 
-class Separator(Item[V]):
+class Separator(ViewItem[V]):
     """Represents a UI Separator.
 
     .. versionadded:: 2.7
@@ -56,10 +80,6 @@ class Separator(Item[V]):
         )
 
     @property
-    def type(self) -> ComponentType:
-        return self._underlying.type
-
-    @property
     def divider(self) -> bool:
         """Whether the separator is a divider. Defaults to ``True``."""
         return self._underlying.divider
@@ -77,12 +97,8 @@ class Separator(Item[V]):
     def spacing(self, value: SeparatorSpacingSize) -> None:
         self._underlying.spacing = value
 
-    @property
-    def width(self) -> int:
-        return 5
-
     def to_component_dict(self) -> SeparatorComponentPayload:
-        return self._underlying.to_dict()
+        return super().to_component_dict()
 
     @classmethod
     def from_component(cls: type[S], component: SeparatorComponent) -> S:
