@@ -65,7 +65,24 @@ from typing import (
     overload,
 )
 
-from .errors import HTTPException, InvalidArgument
+if TYPE_CHECKING:
+    from discord import (
+        Client,
+        VoiceChannel,
+        TextChannel,
+        ForumChannel,
+        StageChannel,
+        CategoryChannel,
+        Thread,
+        Member,
+        User,
+        Guild,
+        Role,
+        GuildEmoji,
+        AppEmoji,
+    )
+
+from .errors import HTTPException, InvalidArgument, InvalidData
 
 try:
     import msgspec
@@ -100,6 +117,7 @@ __all__ = (
     "generate_snowflake",
     "basic_autocomplete",
     "filter_params",
+    "MISSING",
 )
 
 _log = logging.getLogger(__name__)
@@ -567,64 +585,227 @@ def get(iterable: Iterable[T], **attrs: Any) -> T | None:
     return None
 
 
-async def get_or_fetch(obj, attr: str, id: int, *, default: Any = MISSING) -> Any:
-    """|coro|
+_FETCHABLE = TypeVar(
+    "_FETCHABLE",
+    bound="VoiceChannel | TextChannel | ForumChannel | StageChannel | CategoryChannel | Thread | Member | User | Guild | Role | GuildEmoji | AppEmoji",
+)
+_D = TypeVar("_D")
+_Getter = Callable[[Any, int], Any]
+_Fetcher = Callable[[Any, int], Awaitable[Any]]
 
-    Attempts to get an attribute from the object in cache. If it fails, it will attempt to fetch it.
-    If the fetch also fails, an error will be raised.
+
+# TODO: In version 3.0, remove the 'attr' and 'id' arguments.
+#       Also, eliminate the default 'MISSING' value for both 'object_type' and 'object_id'.
+@overload
+async def get_or_fetch(
+    obj: Guild | Client,
+    object_type: type[_FETCHABLE],
+    object_id: Literal[None],
+    default: _D = ...,
+    attr: str = ...,
+    id: int = ...,
+) -> None | _D: ...
+
+
+@overload
+async def get_or_fetch(
+    obj: Guild | Client,
+    object_type: type[_FETCHABLE],
+    object_id: int,
+    default: _D,
+    attr: str = ...,
+    id: int = ...,
+) -> _FETCHABLE | _D: ...
+
+
+@overload
+async def get_or_fetch(
+    obj: Guild | Client,
+    object_type: type[_FETCHABLE],
+    object_id: int,
+    *,
+    attr: str = ...,
+    id: int = ...,
+) -> _FETCHABLE: ...
+
+
+async def get_or_fetch(
+    obj: Guild | Client,
+    object_type: type[_FETCHABLE] = MISSING,
+    object_id: int | None = MISSING,
+    default: _D = MISSING,
+    attr: str = MISSING,
+    id: int = MISSING,
+) -> _FETCHABLE | _D | None:
+    """
+    Shortcut method to get data from an object either by returning the cached version, or if it does not exist, attempting to fetch it from the API.
 
     Parameters
     ----------
-    obj: Any
-        The object to use the get or fetch methods in
-    attr: :class:`str`
-        The attribute to get or fetch. Note the object must have both a ``get_`` and ``fetch_`` method for this attribute.
-    id: :class:`int`
-        The ID of the object
-    default: Any
-        The default value to return if the object is not found, instead of raising an error.
+    obj: :class:`~discord.Guild` | :class:`~discord.Client`
+        The object to operate on.
+
+    object_type: Type[:class:`~discord.VoiceChannel` | :class:`~discord.TextChannel` | :class:`~discord.ForumChannel` | :class:`~discord.StageChannel` | :class:`~discord.CategoryChannel` | :class:`~discord.Thread` | :class:`~discord.User` | :class:`~discord.Guild` | :class:`~discord.Role` | :class:`~discord.Member` | :class:`~discord.GuildEmoji` | :class:`~discord.AppEmoji`]
+        Type of object to fetch or get.
+
+    object_id: :class:`int` | :data:`None`
+        ID of object to get.
+
+    default: Any | :data:`None`
+        The value to return instead of raising if fetching fails.
 
     Returns
     -------
-    Any
-        The object found or the default value.
+    :class:`~discord.VoiceChannel` | :class:`~discord.TextChannel` | :class:`~discord.ForumChannel` | :class:`~discord.StageChannel` | :class:`~discord.CategoryChannel` | :class:`~discord.Thread` | :class:`~discord.User` | :class:`~discord.Guild` | :class:`~discord.Role` | :class:`~discord.Member` | :class:`~discord.GuildEmoji` | :class:`~discord.AppEmoji` | :data:`None`
+        The object if found, or `default` if provided when not found.
+        Returns :data:`None` only if `object_id` is :data:`None` and no `default` is given.
 
     Raises
     ------
-    :exc:`AttributeError`
-        The object is missing a ``get_`` or ``fetch_`` method
+    :exc:`TypeError`
+        Raised when required parameters are missing or invalid types are provided.
+    :exc:`InvalidArgument`
+        Raised when an unsupported or incompatible object type is used.
     :exc:`NotFound`
-        Invalid ID for the object
+        Invalid ID for the object.
     :exc:`HTTPException`
-        An error occurred fetching the object
+        An error occurred fetching the object.
     :exc:`Forbidden`
-        You do not have permission to fetch the object
-
-    Examples
-    --------
-
-    Getting a guild from a guild ID: ::
-
-        guild = await utils.get_or_fetch(client, 'guild', guild_id)
-
-    Getting a channel from the guild. If the channel is not found, return None: ::
-
-        channel = await utils.get_or_fetch(guild, 'channel', channel_id, default=None)
+        You do not have permission to fetch the object.
+    :exc:`InvalidData`
+        Raised when the object resolves to a different guild.
     """
-    getter = getattr(obj, f"get_{attr}")(id)
-    if getter is None:
-        try:
-            getter = await getattr(obj, f"fetch_{attr}")(id)
-        except AttributeError:
-            getter = await getattr(obj, f"_fetch_{attr}")(id)
-            if getter is None:
-                raise ValueError(f"Could not find {attr} with id {id} on {obj}")
-        except (HTTPException, ValueError):
-            if default is not MISSING:
-                return default
-            else:
-                raise
-    return getter
+    from discord import AppEmoji, Client, Guild, Member, Role, User
+
+    if object_id is None:
+        return default if default is not MISSING else None
+
+    # Temporary backward compatibility for 'attr' and 'id'.
+    # This entire if block should be removed in version 3.0.
+    if attr is not MISSING or id is not MISSING or isinstance(object_type, str):
+        warn_deprecated(
+            name="get_or_fetch(obj, attr='type', id=...)",
+            instead="get_or_fetch(obj, object_type=Type, object_id=...)",
+            since="2.7",
+            removed="3.0",
+        )
+
+        deprecated_attr = attr if attr is not MISSING else object_type
+        deprecated_id = id if id is not MISSING else object_id
+
+        if isinstance(deprecated_attr, str):
+            mapped_type = _get_string_to_type_map().get(deprecated_attr.lower())
+            if mapped_type is None:
+                raise InvalidArgument(
+                    f"Unknown type string '{deprecated_attr}' used. Please use a valid class like `discord.Member` instead."
+                )
+            object_type = mapped_type
+        elif isinstance(deprecated_attr, type):
+            object_type = deprecated_attr
+        else:
+            raise TypeError(
+                f"Invalid `attr` or `object_type`: expected a string or class, got {type(deprecated_attr).__name__}."
+            )
+
+        object_id = deprecated_id
+
+    if object_type is MISSING or object_id is MISSING:
+        raise TypeError("required parameters: `object_type` and `object_id`.")
+
+    if isinstance(obj, Guild) and object_type is User:
+        raise InvalidArgument("Guild cannot get_or_fetch User. Use Client instead.")
+    elif isinstance(obj, Client) and object_type is Member:
+        raise InvalidArgument("Client cannot get_or_fetch Member. Use Guild instead.")
+    elif isinstance(obj, Client) and object_type is Role:
+        raise InvalidArgument("Client cannot get_or_fetch Role. Use Guild instead.")
+    elif isinstance(obj, Guild) and object_type is Guild:
+        raise InvalidArgument("Guild cannot get_or_fetch Guild. Use Client instead.")
+    elif isinstance(obj, Guild) and object_type is AppEmoji:
+        raise InvalidArgument("Guild cannot get_or_fetch AppEmoji. Use Client instead.")
+
+    try:
+        getter, fetcher = _get_getter_fetcher_map()[object_type]
+    except KeyError:
+        raise InvalidArgument(
+            f"Class {object_type.__name__} cannot be used with discord.{type(obj).__name__}.get_or_fetch()"
+        )
+
+    result = getter(obj, object_id)
+    if result is not None:
+        return result
+
+    try:
+        return await fetcher(obj, object_id)
+    except (HTTPException, ValueError, InvalidData):
+        if default is not MISSING:
+            return default
+        raise
+
+
+@functools.lru_cache(maxsize=1)
+def _get_string_to_type_map() -> dict[str, type]:
+    """Return a cached map of lowercase strings -> discord types."""
+    from discord import AppEmoji, Guild, Member, Role, User, abc, emoji
+
+    return {
+        "channel": abc.GuildChannel,
+        "member": Member,
+        "user": User,
+        "guild": Guild,
+        "emoji": emoji._EmojiTag,
+        "appemoji": AppEmoji,
+        "role": Role,
+    }
+
+
+@functools.lru_cache(maxsize=1)
+def _get_getter_fetcher_map() -> dict[type, tuple[_Getter, _Fetcher]]:
+    """Return a cached map of type names -> (getter, fetcher) functions."""
+    from discord import Guild, Member, Role, User, abc, emoji
+
+    base_map: dict[type, tuple[_Getter, _Fetcher]] = {
+        Member: (
+            lambda obj, oid: obj.get_member(oid),
+            lambda obj, oid: obj.fetch_member(oid),
+        ),
+        Role: (
+            lambda obj, oid: obj.get_role(oid),
+            lambda obj, oid: obj.fetch_role(oid),
+        ),
+        User: (
+            lambda obj, oid: obj.get_user(oid),
+            lambda obj, oid: obj.fetch_user(oid),
+        ),
+        Guild: (
+            lambda obj, oid: obj.get_guild(oid),
+            lambda obj, oid: obj.fetch_guild(oid),
+        ),
+        emoji._EmojiTag: (
+            lambda obj, oid: obj.get_emoji(oid),
+            lambda obj, oid: obj.fetch_emoji(oid),
+        ),
+        abc.GuildChannel: (
+            lambda obj, oid: obj.get_channel(oid),
+            lambda obj, oid: obj.fetch_channel(oid),
+        ),
+    }
+
+    expanded: dict[type, tuple[_Getter, _Fetcher]] = {}
+    for base, funcs in base_map.items():
+        expanded[base] = funcs
+        for subclass in _all_subclasses(base):
+            if subclass not in expanded:
+                expanded[subclass] = funcs
+
+    return expanded
+
+
+def _all_subclasses(cls: type) -> set[type]:
+    """Recursively collect all subclasses of a class."""
+    subs = set(cls.__subclasses__())
+    for sub in cls.__subclasses__():
+        subs |= _all_subclasses(sub)
+    return subs
 
 
 def _unique(iterable: Iterable[T]) -> list[T]:
